requirements.txt
flake8==7.3.0
mccabe==0.7.0
numpy==2.3.4
pillow==12.0.0
pycodestyle==2.14.0
pyflakes==3.4.0
